refactor: move webhook events utils in utils package
